+++ modules/fmops/sagemaker-jumpstart-fm-endpoint/lib/sagemaker-jumpstart-fm-endpoint-stack.ts
@@ -8,6 +8,7 @@ import {
   SageMakerInstanceType,
   JumpStartSageMakerEndpoint,
 } from "@cdklabs/generative-ai-cdk-constructs";
+import * as cdk_nag from "cdk-nag";
 
 interface SagemakerJumpStartFmEndpointStackProps extends cdk.StackProps {
   projectName?: string;
@@ -37,7 +38,7 @@ export class SagemakerJumpStartFmEndpointStack extends cdk.Stack {
             new PolicyStatement({
               effect: Effect.ALLOW,
               actions: ["s3:GetObject", "s3:ListBucket"],
-              resources: ["arn:aws:s3:::jumpstart-*"],
+              resources: [`arn:${this.partition}:s3:::jumpstart-*`],
             }),
             new PolicyStatement({
               effect: Effect.ALLOW,
@@ -91,5 +92,12 @@ export class SagemakerJumpStartFmEndpointStack extends cdk.Stack {
       role: this.role,
       vpcConfig: vpcConfig,
     });
+
+    cdk_nag.NagSuppressions.addResourceSuppressions(this.role, [
+      {
+        id: "AwsSolutions-IAM5",
+        reason: "Resource access restriced to S3 buckets (with a prefix) and ECR images",
+      },
+    ]);
   }
 }