+++ modules/mlflow/mlflow-fargate/tests/test_stack.py
@@ -3,8 +3,9 @@ import sys
 from unittest import mock
 
 import aws_cdk as cdk
+import cdk_nag
 import pytest
-from aws_cdk.assertions import Template
+from aws_cdk.assertions import Annotations, Match, Template
 
 
 @pytest.fixture(scope="function")
@@ -20,8 +21,8 @@ def stack_defaults():
         yield
 
 
-@pytest.mark.parametrize("use_rds", [False, True])
-def test_synthesize_stack(stack_defaults, use_rds) -> None:
+@pytest.fixture(scope="function")
+def stack(stack_defaults, use_rds: bool) -> cdk.Stack:
     import stack
 
     app = cdk.App()
@@ -38,6 +39,7 @@ def test_synthesize_stack(stack_defaults, use_rds) -> None:
     task_memory_limit_mb = 8 * 1024
     autoscale_max_capacity = 2
     artifacts_bucket_name = "bucket"
+    efs_removal_policy = "DESTROY"
 
     if use_rds:
         rds_settings = {
@@ -49,7 +51,7 @@ def test_synthesize_stack(stack_defaults, use_rds) -> None:
     else:
         rds_settings = None
 
-    stack = stack.MlflowFargateStack(
+    return stack.MlflowFargateStack(
         scope=app,
         id=app_prefix,
         app_prefix=app_prefix,
@@ -65,12 +67,27 @@ def test_synthesize_stack(stack_defaults, use_rds) -> None:
         rds_settings=rds_settings,
         lb_access_logs_bucket_name=None,
         lb_access_logs_bucket_prefix=None,
+        efs_removal_policy=efs_removal_policy,
         env=cdk.Environment(
             account=os.environ["CDK_DEFAULT_ACCOUNT"],
             region=os.environ["CDK_DEFAULT_REGION"],
         ),
     )
 
+
+@pytest.mark.parametrize("use_rds", [False, True])
+def test_synthesize_stack(stack: cdk.Stack, use_rds: bool) -> None:
     template = Template.from_stack(stack)
     template.resource_count_is("AWS::ECS::Cluster", 1)
     template.resource_count_is("AWS::ECS::TaskDefinition", 1)
+
+
+@pytest.mark.parametrize("use_rds", [False, True])
+def test_no_cdk_nag_errors(stack: cdk.Stack, use_rds: bool) -> None:
+    cdk.Aspects.of(stack).add(cdk_nag.AwsSolutionsChecks())
+
+    nag_errors = Annotations.from_stack(stack).find_error(
+        "*",
+        Match.string_like_regexp(r"AwsSolutions-.*"),
+    )
+    assert not nag_errors, f"Found {len(nag_errors)} CDK nag errors"