+++ modules/sagemaker/sagemaker-templates-service-catalog/stack.py
@@ -3,8 +3,9 @@
 
 import importlib
 import os
-from typing import Any, Optional, Tuple
+from typing import Any, List, Optional, Tuple
 
+import cdk_nag
 from aws_cdk import BundlingOptions, BundlingOutput, DockerImage, Stack, Tags
 from aws_cdk import aws_iam as iam
 from aws_cdk import aws_s3_assets as s3_assets
@@ -20,6 +21,19 @@ class ServiceCatalogStack(Stack):
         portfolio_name: str,
         portfolio_owner: str,
         portfolio_access_role_arn: str,
+        dev_vpc_id: str,
+        dev_subnet_ids: List[str],
+        dev_security_group_ids: List[str],
+        pre_prod_account_id: str,
+        pre_prod_region: str,
+        pre_prod_vpc_id: str,
+        pre_prod_subnet_ids: List[str],
+        pre_prod_security_group_ids: List[str],
+        prod_account_id: str,
+        prod_region: str,
+        prod_vpc_id: str,
+        prod_subnet_ids: List[str],
+        prod_security_group_ids: List[str],
         **kwargs: Any,
     ) -> None:
         super().__init__(scope, id, **kwargs)
@@ -60,6 +74,8 @@ class ServiceCatalogStack(Stack):
 
         templates_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "templates")
         for template_name in next(os.walk(templates_dir))[1]:
+            if template_name == "__pycache__":
+                continue
             build_app_asset, deploy_app_asset = self.upload_assets(
                 portfolio_access_role=portfolio_access_role,
                 template_name=template_name,
@@ -71,6 +87,19 @@ class ServiceCatalogStack(Stack):
                 f"{template_name}ProductStack",
                 build_app_asset=build_app_asset,
                 deploy_app_asset=deploy_app_asset,
+                dev_vpc_id=dev_vpc_id,
+                dev_subnet_ids=dev_subnet_ids,
+                dev_security_group_ids=dev_security_group_ids,
+                pre_prod_vpc_id=pre_prod_vpc_id,
+                pre_prod_account_id=pre_prod_account_id,
+                pre_prod_region=pre_prod_region,
+                pre_prod_subnet_ids=pre_prod_subnet_ids,
+                pre_prod_security_group_ids=pre_prod_security_group_ids,
+                prod_vpc_id=prod_vpc_id,
+                prod_account_id=prod_account_id,
+                prod_region=prod_region,
+                prod_subnet_ids=prod_subnet_ids,
+                prod_security_group_ids=prod_security_group_ids,
             )
 
             product_name: str = getattr(product_stack, "TEMPLATE_NAME", template_name)
@@ -96,48 +125,74 @@ class ServiceCatalogStack(Stack):
 
             Tags.of(product).add(key="sagemaker:studio-visibility", value="true")
 
+            cdk_nag.NagSuppressions.add_resource_suppressions(
+                portfolio_access_role,
+                apply_to_children=True,
+                suppressions=[
+                    cdk_nag.NagPackSuppression(
+                        id="AwsSolutions-IAM5",
+                        reason="The role needs wildcard permissions to be able to access template assets in S3",
+                    ),
+                ],
+            )
+            cdk_nag.NagSuppressions.add_resource_suppressions(
+                product_launch_role,
+                suppressions=[
+                    cdk_nag.NagPackSuppression(
+                        id="AwsSolutions-IAM4",
+                        reason=(
+                            "Product launch role needs admin permissions in order to be able "
+                            "to create resources in the AWS account."
+                        ),
+                    ),
+                ],
+            )
+
     def upload_assets(
         self,
         portfolio_access_role: iam.IRole,
         template_name: str,
-    ) -> Tuple[s3_assets.Asset, Optional[s3_assets.Asset]]:
+    ) -> Tuple[Optional[s3_assets.Asset], Optional[s3_assets.Asset]]:
         # Create the build and deployment asset as an output to pass to pipeline stack
         zip_image = DockerImage.from_build("images/zip-image")
 
-        build_app_asset = s3_assets.Asset(
-            self,
-            f"{template_name}BuildAsset",
-            path=f"templates/{template_name}/seed_code/build_app/",
-            bundling=BundlingOptions(
-                image=zip_image,
-                command=[
-                    "sh",
-                    "-c",
-                    """zip -r /asset-output/build_app.zip .""",
-                ],
-                output_type=BundlingOutput.ARCHIVED,
-            ),
-        )
-        build_app_asset.grant_read(grantee=portfolio_access_role)
+        build_path = f"templates/{template_name}/seed_code/build_app/"
+        deploy_path = f"templates/{template_name}/seed_code/deploy_app/"
+        build_app_asset = None
+        deploy_app_asset = None
 
-        # check if there is a deploy_app folder
-        if not os.path.isdir(f"templates/{template_name}/seed_code/deploy_app/"):
-            return build_app_asset, None
+        if os.path.isdir(build_path):
+            build_app_asset = s3_assets.Asset(
+                self,
+                f"{template_name}BuildAsset",
+                path=f"templates/{template_name}/seed_code/build_app/",
+                bundling=BundlingOptions(
+                    image=zip_image,
+                    command=[
+                        "sh",
+                        "-c",
+                        """zip -r /asset-output/build_app.zip .""",
+                    ],
+                    output_type=BundlingOutput.ARCHIVED,
+                ),
+            )
+            build_app_asset.grant_read(grantee=portfolio_access_role)
 
-        deploy_app_asset = s3_assets.Asset(
-            self,
-            f"{template_name}DeployAsset",
-            path=f"templates/{template_name}/seed_code/deploy_app/",
-            bundling=BundlingOptions(
-                image=zip_image,
-                command=[
-                    "sh",
-                    "-c",
-                    """zip -r /asset-output/deploy_app.zip .""",
-                ],
-                output_type=BundlingOutput.ARCHIVED,
-            ),
-        )
-        deploy_app_asset.grant_read(grantee=portfolio_access_role)
+        if os.path.isdir(deploy_path):
+            deploy_app_asset = s3_assets.Asset(
+                self,
+                f"{template_name}DeployAsset",
+                path=f"templates/{template_name}/seed_code/deploy_app/",
+                bundling=BundlingOptions(
+                    image=zip_image,
+                    command=[
+                        "sh",
+                        "-c",
+                        """zip -r /asset-output/deploy_app.zip .""",
+                    ],
+                    output_type=BundlingOutput.ARCHIVED,
+                ),
+            )
+            deploy_app_asset.grant_read(grantee=portfolio_access_role)
 
         return build_app_asset, deploy_app_asset