+++ modules/sagemaker/sagemaker-templates-service-catalog/tests/test_stack.py
@@ -5,8 +5,9 @@ import os
 import sys
 
 import aws_cdk as cdk
+import cdk_nag
 import pytest
-from aws_cdk.assertions import Template
+from aws_cdk.assertions import Annotations, Match, Template
 
 
 @pytest.fixture(scope="function")
@@ -20,7 +21,8 @@ def stack_defaults():
         del sys.modules["stack"]
 
 
-def test_synthesize_stack(stack_defaults):
+@pytest.fixture(scope="function")
+def stack(stack_defaults) -> cdk.Stack:
     import stack
 
     app = cdk.App()
@@ -30,20 +32,58 @@ def test_synthesize_stack(stack_defaults):
     portfolio_name = "portfolio"
     portfolio_owner = "owner"
     portfolio_access_role_arn = "arn:aws:iam::xxxxxxxxxxxx:role/role"
+    dev_vpc_id = "vpc"
+    dev_subnet_ids = ["sub"]
+    dev_security_group_ids = ["sg"]
+    pre_prod_account_id = "pre_prod_account_id"
+    pre_prod_region = "us-east-1"
+    pre_prod_vpc_id = "vpc"
+    pre_prod_subnet_ids = ["sub"]
+    pre_prod_security_group_ids = ["sg"]
+    prod_account_id = "prod_account_id"
+    prod_region = "us-east-1"
+    prod_vpc_id = "vpc"
+    prod_subnet_ids = ["sub"]
+    prod_security_group_ids = ["sg"]
 
-    stack = stack.ServiceCatalogStack(
+    return stack.ServiceCatalogStack(
         app,
         f"{project_name}-{dep_name}-{mod_name}",
         portfolio_name=portfolio_name,
         portfolio_owner=portfolio_owner,
         portfolio_access_role_arn=portfolio_access_role_arn,
+        dev_vpc_id=dev_vpc_id,
+        dev_subnet_ids=dev_subnet_ids,
+        dev_security_group_ids=dev_security_group_ids,
+        pre_prod_account_id=pre_prod_account_id,
+        pre_prod_region=pre_prod_region,
+        pre_prod_vpc_id=pre_prod_vpc_id,
+        pre_prod_subnet_ids=pre_prod_subnet_ids,
+        pre_prod_security_group_ids=pre_prod_security_group_ids,
+        prod_account_id=prod_account_id,
+        prod_region=prod_region,
+        prod_vpc_id=prod_vpc_id,
+        prod_subnet_ids=prod_subnet_ids,
+        prod_security_group_ids=prod_security_group_ids,
         env=cdk.Environment(
             account=os.environ["CDK_DEFAULT_ACCOUNT"],
             region=os.environ["CDK_DEFAULT_REGION"],
         ),
     )
 
+
+def test_synthesize_stack(stack: cdk.Stack) -> None:
     template = Template.from_stack(stack)
 
     template.resource_count_is("AWS::ServiceCatalog::Portfolio", 1)
-    template.resource_count_is("AWS::ServiceCatalog::CloudFormationProduct", 1)
+    template.resource_count_is("AWS::ServiceCatalog::CloudFormationProduct", 3)
+
+
+def test_no_cdk_nag_errors(stack: cdk.Stack) -> None:
+    cdk.Aspects.of(stack).add(cdk_nag.AwsSolutionsChecks())
+
+    nag_errors = Annotations.from_stack(stack).find_error(
+        "*",
+        Match.string_like_regexp(r"AwsSolutions-.*"),
+    )
+    assert not nag_errors, f"Found {len(nag_errors)} CDK nag errors"