+++ scripts/fix.sh
@@ -46,7 +46,7 @@ FIX_PATH=`pwd`
 echo "Fixing: ${FIX_PATH}, Language: ${LANGUAGE}"
 
 if [[ $LANGUAGE == "python" ]]; then
-    echo "Running isort, black"
+    echo "Running ruff"
     ruff format .
     ruff check --fix .
 elif [[ $LANGUAGE == "typescript" ]]; then